Split front-end code into smaller files
Browsers don't have a concept of "importing" javascript into other files, so create a global variable to store anything which need to be "exported" from a file for use in other files. Make index.js simply load all the files in appropriate order. (This is not a scalable solution because it requires us to manually define the dependencies and loading order, but it's lighter-weight than webpack or something where you transpile imported code)