Add a form for Username
It will load first, and when the user provides a name it will load the regular search form with any previous searches stored in it (not currently functional, though).

I'm a little worried about url encoding the name, since it becomes a cookie and is therefore encoded. It seems to be working without it, perhaps the cookie-parser library is helping.